Protectile test in one example
